refactor all factories and add makefile
