+++ interface/boosted_factory.h
@@ -7,9 +7,10 @@
 #include <memory>
 #include <string>
 #include <vector>
-#include "./boosted.h"
+#include "./util.h"
 #include "TTree.h"
 
+class Boosted;
 typedef std::vector<Boosted> VBoosted;
 
 class Boosted_Factory {
@@ -33,11 +34,9 @@ class Boosted_Factory {
     std::vector<Bool_t> *boostedTaupfTausDiscriminationByDecayModeFinding, *boostedTaupfTausDiscriminationByDecayModeFindingNewDMs,
         *boostedTauByMVA6VLooseElectronRejection, *boostedTauByMVA6LooseElectronRejection, *boostedTauByMVA6MediumElectronRejection,
         *boostedTauByMVA6TightElectronRejection, *boostedTauByMVA6VTightElectronRejection, *boostedTauByLooseMuonRejection3,
-        *boostedTauByTightMuonRejection3, *boostedTauLeadChargedHadronExists;
+        *boostedTauByTightMuonRejection3;
 
-    std::vector<Int_t> *boostedTauDecayMode, *boostedTauNumSignalPFChargedHadrCands, *boostedTauNumSignalPFNeutrHadrCands,
-        *boostedTauNumSignalPFGammaCands, *boostedTauNumSignalPFCands, *boostedTauNumIsolationPFChargedHadrCands,
-        *boostedTauNumIsolationPFNeutrHadrCands, *boostedTauNumIsolationPFGammaCands, *boostedTauNumIsolationPFCands;
+    std::vector<Int_t> *boostedTauDecayMode;
 };
 
 Boosted_Factory::Boosted_Factory(TTree *tree, std::string isoType = "IsolationMVArun2v2DBoldDMwLT")
@@ -60,16 +59,7 @@ Boosted_Factory::Boosted_Factory(TTree *tree, std::string isoType = "IsolationMV
       boostedTauByMVA6VTightElectronRejection(nullptr),
       boostedTauByLooseMuonRejection3(nullptr),
       boostedTauByTightMuonRejection3(nullptr),
-      boostedTauLeadChargedHadronExists(nullptr),
       boostedTauDecayMode(nullptr),
-      boostedTauNumSignalPFChargedHadrCands(nullptr),
-      boostedTauNumSignalPFNeutrHadrCands(nullptr),
-      boostedTauNumSignalPFGammaCands(nullptr),
-      boostedTauNumSignalPFCands(nullptr),
-      boostedTauNumIsolationPFChargedHadrCands(nullptr),
-      boostedTauNumIsolationPFNeutrHadrCands(nullptr),
-      boostedTauNumIsolationPFGammaCands(nullptr),
-      boostedTauNumIsolationPFCands(nullptr),
       boostedTauCharge(nullptr),
       boostedTaudz(nullptr),
       boostedTaudxy(nullptr) {
@@ -78,6 +68,7 @@ Boosted_Factory::Boosted_Factory(TTree *tree, std::string isoType = "IsolationMV
     tree->SetBranchAddress("boostedTauEta", &boostedTauEta);
     tree->SetBranchAddress("boostedTauPhi", &boostedTauPhi);
     tree->SetBranchAddress("boostedTauMass", &boostedTauMass);
+    // only 1-type of isolation available, for now
     tree->SetBranchAddress("boostedTauByIsolationMVArun2v1DBoldDMwLTraw", &iso);
     tree->SetBranchAddress("boostedTauByVLooseIsolationMVArun2v1DBoldDMwLT", &pass_vloose_iso);
     tree->SetBranchAddress("boostedTauByLooseIsolationMVArun2v1DBoldDMwLT", &pass_loose_iso);
@@ -93,16 +84,7 @@ Boosted_Factory::Boosted_Factory(TTree *tree, std::string isoType = "IsolationMV
     tree->SetBranchAddress("boostedTauByMVA6VTightElectronRejection", &boostedTauByMVA6VTightElectronRejection);
     tree->SetBranchAddress("boostedTauByLooseMuonRejection3", &boostedTauByLooseMuonRejection3);
     tree->SetBranchAddress("boostedTauByTightMuonRejection3", &boostedTauByTightMuonRejection3);
-    tree->SetBranchAddress("boostedTauLeadChargedHadronExists", &boostedTauLeadChargedHadronExists);
     tree->SetBranchAddress("boostedTauDecayMode", &boostedTauDecayMode);
-    tree->SetBranchAddress("boostedTauNumSignalPFChargedHadrCands", &boostedTauNumSignalPFChargedHadrCands);
-    tree->SetBranchAddress("boostedTauNumSignalPFNeutrHadrCands", &boostedTauNumSignalPFNeutrHadrCands);
-    tree->SetBranchAddress("boostedTauNumSignalPFGammaCands", &boostedTauNumSignalPFGammaCands);
-    tree->SetBranchAddress("boostedTauNumSignalPFCands", &boostedTauNumSignalPFCands);
-    tree->SetBranchAddress("boostedTauNumIsolationPFChargedHadrCands", &boostedTauNumIsolationPFChargedHadrCands);
-    tree->SetBranchAddress("boostedTauNumIsolationPFNeutrHadrCands", &boostedTauNumIsolationPFNeutrHadrCands);
-    tree->SetBranchAddress("boostedTauNumIsolationPFGammaCands", &boostedTauNumIsolationPFGammaCands);
-    tree->SetBranchAddress("boostedTauNumIsolationPFCands", &boostedTauNumIsolationPFCands);
     tree->SetBranchAddress("boostedTauCharge", &boostedTauCharge);
     tree->SetBranchAddress("boostedTaudz", &boostedTaudz);
     tree->SetBranchAddress("boostedTaudxy", &boostedTaudxy);
@@ -115,6 +97,7 @@ void Boosted_Factory::Run_Factory() {
             continue;
         }
         auto boosted = Boosted(boostedTauPt->at(i), boostedTauEta->at(i), boostedTauPhi->at(i), boostedTauMass->at(i));
+        boosted.Iso = iso->at(i);
         boosted.pass_vloose_iso = pass_vloose_iso->at(i);
         boosted.pass_loose_iso = pass_loose_iso->at(i);
         boosted.pass_medium_iso = pass_medium_iso->at(i);
@@ -129,16 +112,7 @@ void Boosted_Factory::Run_Factory() {
         boosted.ByMVA6VTightElectronRejection = boostedTauByMVA6VTightElectronRejection->at(i);
         boosted.ByLooseMuonRejection3 = boostedTauByLooseMuonRejection3->at(i);
         boosted.ByTightMuonRejection3 = boostedTauByTightMuonRejection3->at(i);
-        boosted.LeadChargedHadronExists = boostedTauLeadChargedHadronExists->at(i);
         boosted.DecayMode = boostedTauDecayMode->at(i);
-        boosted.NumSignalPFChargedHadrCands = boostedTauNumSignalPFChargedHadrCands->at(i);
-        boosted.NumSignalPFNeutrHadrCands = boostedTauNumSignalPFNeutrHadrCands->at(i);
-        boosted.NumSignalPFGammaCands = boostedTauNumSignalPFGammaCands->at(i);
-        boosted.NumSignalPFCands = boostedTauNumSignalPFCands->at(i);
-        boosted.NumIsolationPFChargedHadrCands = boostedTauNumIsolationPFChargedHadrCands->at(i);
-        boosted.NumIsolationPFNeutrHadrCands = boostedTauNumIsolationPFNeutrHadrCands->at(i);
-        boosted.NumIsolationPFGammaCands = boostedTauNumIsolationPFGammaCands->at(i);
-        boosted.NumIsolationPFCands = boostedTauNumIsolationPFCands->at(i);
         boosted.Charge = boostedTauCharge->at(i);
         boosted.dz = boostedTaudz->at(i);
         boosted.dxy = boostedTaudxy->at(i);
@@ -148,6 +122,88 @@ void Boosted_Factory::Run_Factory() {
     // sort by pT
     std::sort(boosteds.begin(), boosteds.end(), [](Boosted &p1, Boosted &p2) -> bool { return p1.getPt() > p2.getPt(); });
     nGoodTaus = boosteds.size();
+}
+
+class Boosted {
+    friend class Boosted_Factory;
+
+   public:
+    Boosted(Float_t _pt, Float_t _eta, Float_t _phi, Float_t _m) { this->p4.SetPtEtaPhiM(_pt, _eta, _phi, _m); }
+    Boosted() { this->p4.SetPtEtaPhiE(0, 0, 0, 0); }
+
+    // getters
+    TLorentzVector getP4() { return p4; }
+    Float_t getPt() { return p4.Pt(); }
+    Float_t getEta() { return p4.Eta(); }
+    Float_t getPhi() { return p4.Phi(); }
+    Float_t getMass() { return p4.M(); }
+    Float_t getIsoRaw() { return Iso; }
+    Bool_t getIso(working_point);
+    Bool_t getDiscByDM(bool);
+    Bool_t getEleRejection(working_point);
+    Bool_t getMuRejection(working_point);
+    Int_t getDecayMode() { return DecayMode; }
+    Float_t getCharge() { return Charge; }
+    Float_t getDZ() { return dz; }
+    Float_t getDXY() { return dxy; }
+
+   private:
+    TLorentzVector p4;
+    Bool_t pass_vloose_iso, pass_loose_iso, pass_medium_iso, pass_tight_iso, pass_vtight_iso;
+
+    Bool_t pfTausDiscriminationByDecayModeFinding, pfTausDiscriminationByDecayModeFindingNewDMs, ByMVA6VLooseElectronRejection,
+        ByMVA6LooseElectronRejection, ByMVA6MediumElectronRejection, ByMVA6TightElectronRejection, ByMVA6VTightElectronRejection,
+        ByLooseMuonRejection3, ByTightMuonRejection3;
+
+    Int_t DecayMode;
+
+    Float_t Charge, Mass, dz, dxy, Energy, Iso;
+};
+
+Bool_t Boosted::getIso(working_point wp) {
+    if (wp == vloose) {
+        return pass_vloose_iso;
+    } else if (wp == loose) {
+        return pass_loose_iso;
+    } else if (wp == medium) {
+        return pass_medium_iso;
+    } else if (wp == tight) {
+        return pass_tight_iso;
+    } else if (wp == vtight) {
+        return pass_vtight_iso;
+    }
+    throw std::invalid_argument("Must use one of the provided tau isolation WPs");
+}
+
+Bool_t Boosted::getDiscByDM(bool newDM = false) {
+    if (newDM) {
+        return pfTausDiscriminationByDecayModeFindingNewDMs;
+    }
+    return pfTausDiscriminationByDecayModeFinding;
+}
+
+Bool_t Boosted::getEleRejection(working_point wp) {
+    if (wp == vloose) {
+        return ByMVA6VLooseElectronRejection;
+    } else if (wp == loose) {
+        return ByMVA6LooseElectronRejection;
+    } else if (wp == medium) {
+        return ByMVA6MediumElectronRejection;
+    } else if (wp == tight) {
+        return ByMVA6TightElectronRejection;
+    } else if (wp == vtight) {
+        return ByMVA6VTightElectronRejection;
+    }
+    throw std::invalid_argument("Must use one of the provided electron rejection WPs");
+}
+
+Bool_t Boosted::getMuRejection(working_point wp) {
+    if (wp == loose) {
+        return ByLooseMuonRejection3;
+    } else if (wp == tight) {
+        return ByTightMuonRejection3;
+    }
+    throw std::invalid_argument("Muon rejection working point was neither Tight nor Loose");
 }
 
 #endif  // INTERFACE_BOOSTED_FACTORY_H_
