+++ interface/tau_factory.h
@@ -7,37 +7,36 @@
 #include <memory>
 #include <string>
 #include <vector>
-#include "./tau.h"
+#include "./util.h"
+#include "TLorentzVector.h"
 #include "TTree.h"
 
+class Tau;
 typedef std::vector<Tau> VTau;
 
 class Tau_Factory {
- public:
-  explicit Tau_Factory(TTree *, std::string);
-  void Run_Factory();
+   public:
+    explicit Tau_Factory(TTree *, std::string);
+    void Run_Factory();
 
-  // getters
-  Int_t getNTotalTau() { return nTau; }
-  Int_t getNGoodTau() { return nGoodTau; }
-  std::shared_ptr<VTau> getTaus() { return std::make_shared<VTau>(taus); }
+    // getters
+    Int_t getNTotalTau() { return nTau; }
+    Int_t getNGoodTau() { return nGoodTau; }
+    std::shared_ptr<VTau> getTaus() { return std::make_shared<VTau>(taus); }
 
- private:
-  Int_t nTau, nGoodTau;
-  VTau taus;
+   private:
+    Int_t nTau, nGoodTau;
+    VTau taus;
 
-  std::vector<Bool_t> *taupfTausDiscriminationByDecayModeFinding, *taupfTausDiscriminationByDecayModeFindingNewDMs, *tauByMVA6VLooseElectronRejection, *tauByMVA6LooseElectronRejection,
-      *tauByMVA6MediumElectronRejection, *tauByMVA6TightElectronRejection, *tauByMVA6VTightElectronRejection, *tauByLooseMuonRejection3, *tauByTightMuonRejection3,
-      *tauByLooseCombinedIsolationDeltaBetaCorr3Hits, *tauByMediumCombinedIsolationDeltaBetaCorr3Hits, *tauByTightCombinedIsolationDeltaBetaCorr3Hits, *tauLeadChargedHadronExists;
+    std::vector<Bool_t> *taupfTausDiscriminationByDecayModeFinding, *taupfTausDiscriminationByDecayModeFindingNewDMs,
+        *tauByMVA6VLooseElectronRejection, *tauByMVA6LooseElectronRejection, *tauByMVA6MediumElectronRejection, *tauByMVA6TightElectronRejection,
+        *tauByMVA6VTightElectronRejection, *tauByLooseMuonRejection3, *tauByTightMuonRejection3;
 
-  std::vector<Bool_t> *pass_vloose_iso, *pass_loose_iso, *pass_medium_iso, *pass_tight_iso, *pass_vtight_iso;
+    std::vector<Bool_t> *pass_vloose_iso, *pass_loose_iso, *pass_medium_iso, *pass_tight_iso, *pass_vtight_iso;
 
-  std::vector<Int_t> *tauDecayMode, *tauNumSignalPFChargedHadrCands, *tauNumSignalPFNeutrHadrCands, *tauNumSignalPFGammaCands, *tauNumSignalPFCands, *tauNumIsolationPFChargedHadrCands,
-      *tauNumIsolationPFNeutrHadrCands, *tauNumIsolationPFGammaCands, *tauNumIsolationPFCands;
+    std::vector<Int_t> *tauDecayMode;
 
-  std::vector<Float_t> *tauCombinedIsolationDeltaBetaCorrRaw3Hits, *tauEta, *tauPhi, *tauPt, *tauEt, *tauCharge, *tauP, *tauPx, *tauPy, *tauPz, *tauVz, *tauEnergy, *tauMass, *tauDxy,
-      *tauZImpact, *tauLeadChargedHadronEta, *tauLeadChargedHadronPhi, *tauLeadChargedHadronPt, *tauChargedIsoPtSum, *tauNeutralIsoPtSum, *tauPuCorrPtSum, *taufootprintCorrection,
-      *tauphotonPtSumOutsideSignalCone, *taudz, *taudxy, *iso;
+    std::vector<Float_t> *tauEta, *tauPhi, *tauPt, *tauCharge, *tauP, *tauPx, *tauPy, *tauPz, *tauMass, *tauDxy, *taudz, *taudxy, *iso;
 };
 
 Tau_Factory::Tau_Factory(TTree *tree, std::string isoType = "IsolationMVArun2v1DBoldDMwLT")
@@ -60,150 +59,161 @@ Tau_Factory::Tau_Factory(TTree *tree, std::string isoType = "IsolationMVArun2v1D
       tauByMVA6VTightElectronRejection(nullptr),
       tauByLooseMuonRejection3(nullptr),
       tauByTightMuonRejection3(nullptr),
-      tauByLooseCombinedIsolationDeltaBetaCorr3Hits(nullptr),
-      tauByMediumCombinedIsolationDeltaBetaCorr3Hits(nullptr),
-      tauByTightCombinedIsolationDeltaBetaCorr3Hits(nullptr),
-      tauLeadChargedHadronExists(nullptr),
       tauDecayMode(nullptr),
-      tauNumSignalPFChargedHadrCands(nullptr),
-      tauNumSignalPFNeutrHadrCands(nullptr),
-      tauNumSignalPFGammaCands(nullptr),
-      tauNumSignalPFCands(nullptr),
-      tauNumIsolationPFChargedHadrCands(nullptr),
-      tauNumIsolationPFNeutrHadrCands(nullptr),
-      tauNumIsolationPFGammaCands(nullptr),
-      tauNumIsolationPFCands(nullptr),
-      tauCombinedIsolationDeltaBetaCorrRaw3Hits(nullptr),
       tauCharge(nullptr),
       tauP(nullptr),
       tauPx(nullptr),
       tauPy(nullptr),
       tauPz(nullptr),
-      tauVz(nullptr),
-      tauEnergy(nullptr),
       tauDxy(nullptr),
-      tauZImpact(nullptr),
-      tauLeadChargedHadronEta(nullptr),
-      tauLeadChargedHadronPhi(nullptr),
-      tauLeadChargedHadronPt(nullptr),
-      tauChargedIsoPtSum(nullptr),
-      tauNeutralIsoPtSum(nullptr),
-      tauPuCorrPtSum(nullptr),
-      taufootprintCorrection(nullptr),
-      tauphotonPtSumOutsideSignalCone(nullptr),
       taudz(nullptr),
       taudxy(nullptr) {
-  tree->SetBranchAddress("nTau", &nTau);
-  tree->SetBranchAddress("tauPt", &tauPt);
-  tree->SetBranchAddress("tauEta", &tauEta);
-  tree->SetBranchAddress("tauPhi", &tauPhi);
-  tree->SetBranchAddress("tauMass", &tauMass);
-  tree->SetBranchAddress(("tauBy"+isoType+"raw").c_str(), &iso);
-  tree->SetBranchAddress(("tauByVLoose"+isoType).c_str(), &pass_vloose_iso);
-  tree->SetBranchAddress(("tauByLoose"+isoType).c_str(), &pass_loose_iso);
-  tree->SetBranchAddress(("tauByMedium"+isoType).c_str(), &pass_medium_iso);
-  tree->SetBranchAddress(("tauByTight"+isoType).c_str(), &pass_tight_iso);
-  tree->SetBranchAddress(("tauByVTight"+isoType).c_str(), &pass_vtight_iso);
-  tree->SetBranchAddress("taupfTausDiscriminationByDecayModeFinding", &taupfTausDiscriminationByDecayModeFinding);
-  tree->SetBranchAddress("taupfTausDiscriminationByDecayModeFindingNewDMs", &taupfTausDiscriminationByDecayModeFindingNewDMs);
-  tree->SetBranchAddress("tauByMVA6VLooseElectronRejection", &tauByMVA6VLooseElectronRejection);
-  tree->SetBranchAddress("tauByMVA6LooseElectronRejection", &tauByMVA6LooseElectronRejection);
-  tree->SetBranchAddress("tauByMVA6MediumElectronRejection", &tauByMVA6MediumElectronRejection);
-  tree->SetBranchAddress("tauByMVA6TightElectronRejection", &tauByMVA6TightElectronRejection);
-  tree->SetBranchAddress("tauByMVA6VTightElectronRejection", &tauByMVA6VTightElectronRejection);
-  tree->SetBranchAddress("tauByLooseMuonRejection3", &tauByLooseMuonRejection3);
-  tree->SetBranchAddress("tauByTightMuonRejection3", &tauByTightMuonRejection3);
-  tree->SetBranchAddress("tauByLooseCombinedIsolationDeltaBetaCorr3Hits", &tauByLooseCombinedIsolationDeltaBetaCorr3Hits);
-  tree->SetBranchAddress("tauByMediumCombinedIsolationDeltaBetaCorr3Hits", &tauByMediumCombinedIsolationDeltaBetaCorr3Hits);
-  tree->SetBranchAddress("tauByTightCombinedIsolationDeltaBetaCorr3Hits", &tauByTightCombinedIsolationDeltaBetaCorr3Hits);
-  tree->SetBranchAddress("tauLeadChargedHadronExists", &tauLeadChargedHadronExists);
-  tree->SetBranchAddress("tauDecayMode", &tauDecayMode);
-  tree->SetBranchAddress("tauNumSignalPFChargedHadrCands", &tauNumSignalPFChargedHadrCands);
-  tree->SetBranchAddress("tauNumSignalPFNeutrHadrCands", &tauNumSignalPFNeutrHadrCands);
-  tree->SetBranchAddress("tauNumSignalPFGammaCands", &tauNumSignalPFGammaCands);
-  tree->SetBranchAddress("tauNumSignalPFCands", &tauNumSignalPFCands);
-  tree->SetBranchAddress("tauNumIsolationPFChargedHadrCands", &tauNumIsolationPFChargedHadrCands);
-  tree->SetBranchAddress("tauNumIsolationPFNeutrHadrCands", &tauNumIsolationPFNeutrHadrCands);
-  tree->SetBranchAddress("tauNumIsolationPFGammaCands", &tauNumIsolationPFGammaCands);
-  tree->SetBranchAddress("tauNumIsolationPFCands", &tauNumIsolationPFCands);
-  tree->SetBranchAddress("tauCombinedIsolationDeltaBetaCorrRaw3Hits", &tauCombinedIsolationDeltaBetaCorrRaw3Hits);
-  tree->SetBranchAddress("tauCharge", &tauCharge);
-  tree->SetBranchAddress("tauP", &tauP);
-  tree->SetBranchAddress("tauPx", &tauPx);
-  tree->SetBranchAddress("tauPy", &tauPy);
-  tree->SetBranchAddress("tauPz", &tauPz);
-  tree->SetBranchAddress("tauVz", &tauVz);
-  tree->SetBranchAddress("tauEnergy", &tauEnergy);
-  tree->SetBranchAddress("tauDxy", &tauDxy);
-  tree->SetBranchAddress("tauZImpact", &tauZImpact);
-  tree->SetBranchAddress("tauLeadChargedHadronEta", &tauLeadChargedHadronEta);
-  tree->SetBranchAddress("tauLeadChargedHadronPhi", &tauLeadChargedHadronPhi);
-  tree->SetBranchAddress("tauLeadChargedHadronPt", &tauLeadChargedHadronPt);
-  tree->SetBranchAddress("tauChargedIsoPtSum", &tauChargedIsoPtSum);
-  tree->SetBranchAddress("tauNeutralIsoPtSum", &tauNeutralIsoPtSum);
-  tree->SetBranchAddress("tauPuCorrPtSum", &tauPuCorrPtSum);
-  tree->SetBranchAddress("taufootprintCorrection", &taufootprintCorrection);
-  tree->SetBranchAddress("tauphotonPtSumOutsideSignalCone", &tauphotonPtSumOutsideSignalCone);
-  tree->SetBranchAddress("taudz", &taudz);
-  tree->SetBranchAddress("taudxy", &taudxy);
+    tree->SetBranchAddress("nTau", &nTau);
+    tree->SetBranchAddress("tauPt", &tauPt);
+    tree->SetBranchAddress("tauEta", &tauEta);
+    tree->SetBranchAddress("tauPhi", &tauPhi);
+    tree->SetBranchAddress("tauMass", &tauMass);
+    tree->SetBranchAddress(("tauBy" + isoType + "raw").c_str(), &iso);
+    tree->SetBranchAddress(("tauByVLoose" + isoType).c_str(), &pass_vloose_iso);
+    tree->SetBranchAddress(("tauByLoose" + isoType).c_str(), &pass_loose_iso);
+    tree->SetBranchAddress(("tauByMedium" + isoType).c_str(), &pass_medium_iso);
+    tree->SetBranchAddress(("tauByTight" + isoType).c_str(), &pass_tight_iso);
+    tree->SetBranchAddress(("tauByVTight" + isoType).c_str(), &pass_vtight_iso);
+    tree->SetBranchAddress("taupfTausDiscriminationByDecayModeFinding", &taupfTausDiscriminationByDecayModeFinding);
+    tree->SetBranchAddress("taupfTausDiscriminationByDecayModeFindingNewDMs", &taupfTausDiscriminationByDecayModeFindingNewDMs);
+    tree->SetBranchAddress("tauByMVA6VLooseElectronRejection", &tauByMVA6VLooseElectronRejection);
+    tree->SetBranchAddress("tauByMVA6LooseElectronRejection", &tauByMVA6LooseElectronRejection);
+    tree->SetBranchAddress("tauByMVA6MediumElectronRejection", &tauByMVA6MediumElectronRejection);
+    tree->SetBranchAddress("tauByMVA6TightElectronRejection", &tauByMVA6TightElectronRejection);
+    tree->SetBranchAddress("tauByMVA6VTightElectronRejection", &tauByMVA6VTightElectronRejection);
+    tree->SetBranchAddress("tauByLooseMuonRejection3", &tauByLooseMuonRejection3);
+    tree->SetBranchAddress("tauByTightMuonRejection3", &tauByTightMuonRejection3);
+    tree->SetBranchAddress("tauDecayMode", &tauDecayMode);
+    tree->SetBranchAddress("tauCharge", &tauCharge);
+    tree->SetBranchAddress("tauP", &tauP);
+    tree->SetBranchAddress("tauPx", &tauPx);
+    tree->SetBranchAddress("tauPy", &tauPy);
+    tree->SetBranchAddress("tauPz", &tauPz);
+    tree->SetBranchAddress("tauDxy", &tauDxy);
+    tree->SetBranchAddress("taudz", &taudz);
+    tree->SetBranchAddress("taudxy", &taudxy);
 }
 
 void Tau_Factory::Run_Factory() {
-  taus.clear();
-  for (auto i = 0; i < nTau; i++) {
-    // baseline/default selection
-    if (tauPt->at(i) > 20 && tauEta->at(i) < 2.3 && pass_vloose_iso->at(i)) {
-      auto tau = Tau(tauPt->at(i), tauEta->at(i), tauPhi->at(i));  // build the tau
-      tau.pass_vloose_iso = pass_vloose_iso->at(i);
-      tau.pass_loose_iso = pass_loose_iso->at(i);
-      tau.pass_medium_iso = pass_medium_iso->at(i);
-      tau.pass_tight_iso = pass_tight_iso->at(i);
-      tau.pass_vtight_iso = pass_vtight_iso->at(i);
-      tau.pfTausDiscriminationByDecayModeFinding = taupfTausDiscriminationByDecayModeFinding->at(i);
-      tau.pfTausDiscriminationByDecayModeFindingNewDMs = taupfTausDiscriminationByDecayModeFindingNewDMs->at(i);
-      tau.ByMVA6VLooseElectronRejection = tauByMVA6VLooseElectronRejection->at(i);
-      tau.ByMVA6LooseElectronRejection = tauByMVA6LooseElectronRejection->at(i);
-      tau.ByMVA6MediumElectronRejection = tauByMVA6MediumElectronRejection->at(i);
-      tau.ByMVA6TightElectronRejection = tauByMVA6TightElectronRejection->at(i);
-      tau.ByMVA6VTightElectronRejection = tauByMVA6VTightElectronRejection->at(i);
-      tau.ByLooseMuonRejection3 = tauByLooseMuonRejection3->at(i);
-      tau.ByTightMuonRejection3 = tauByTightMuonRejection3->at(i);
-      tau.ByLooseCombinedIsolationDeltaBetaCorr3Hits = tauByLooseCombinedIsolationDeltaBetaCorr3Hits->at(i);
-      tau.ByMediumCombinedIsolationDeltaBetaCorr3Hits = tauByMediumCombinedIsolationDeltaBetaCorr3Hits->at(i);
-      tau.ByTightCombinedIsolationDeltaBetaCorr3Hits = tauByTightCombinedIsolationDeltaBetaCorr3Hits->at(i);
-      tau.LeadChargedHadronExists = tauLeadChargedHadronExists->at(i);
-      tau.DecayMode = tauDecayMode->at(i);
-      tau.NumSignalPFChargedHadrCands = tauNumSignalPFChargedHadrCands->at(i);
-      tau.NumSignalPFNeutrHadrCands = tauNumSignalPFNeutrHadrCands->at(i);
-      tau.NumSignalPFGammaCands = tauNumSignalPFGammaCands->at(i);
-      tau.NumSignalPFCands = tauNumSignalPFCands->at(i);
-      tau.NumIsolationPFChargedHadrCands = tauNumIsolationPFChargedHadrCands->at(i);
-      tau.NumIsolationPFNeutrHadrCands = tauNumIsolationPFNeutrHadrCands->at(i);
-      tau.NumIsolationPFGammaCands = tauNumIsolationPFGammaCands->at(i);
-      tau.NumIsolationPFCands = tauNumIsolationPFCands->at(i);
-      tau.CombinedIsolationDeltaBetaCorrRaw3Hits = tauCombinedIsolationDeltaBetaCorrRaw3Hits->at(i);
-      tau.Charge = tauCharge->at(i);
-      tau.Vz = tauVz->at(i);
-      tau.Energy = tauEnergy->at(i);
-      tau.Dxy = tauDxy->at(i);
-      tau.ZImpact = tauZImpact->at(i);
-      tau.LeadChargedHadronEta = tauLeadChargedHadronEta->at(i);
-      tau.LeadChargedHadronPhi = tauLeadChargedHadronPhi->at(i);
-      tau.LeadChargedHadronPt = tauLeadChargedHadronPt->at(i);
-      tau.ChargedIsoPtSum = tauChargedIsoPtSum->at(i);
-      tau.NeutralIsoPtSum = tauNeutralIsoPtSum->at(i);
-      tau.PuCorrPtSum = tauPuCorrPtSum->at(i);
-      tau.footprintCorrection = taufootprintCorrection->at(i);
-      tau.photonPtSumOutsideSignalCone = tauphotonPtSumOutsideSignalCone->at(i);
-      tau.dz = taudz->at(i);
-      tau.dxy = taudxy->at(i);
-      taus.push_back(tau);
+    taus.clear();
+    for (auto i = 0; i < nTau; i++) {
+        // baseline/default selection
+        if (tauPt->at(i) > 20 && tauEta->at(i) < 2.3 && pass_vloose_iso->at(i)) {
+            auto tau = Tau(tauPt->at(i), tauEta->at(i), tauPhi->at(i), tauMass->at(i));  // build the tau
+            tau.pass_vloose_iso = pass_vloose_iso->at(i);
+            tau.pass_loose_iso = pass_loose_iso->at(i);
+            tau.pass_medium_iso = pass_medium_iso->at(i);
+            tau.pass_tight_iso = pass_tight_iso->at(i);
+            tau.pass_vtight_iso = pass_vtight_iso->at(i);
+            tau.pfTausDiscriminationByDecayModeFinding = taupfTausDiscriminationByDecayModeFinding->at(i);
+            tau.pfTausDiscriminationByDecayModeFindingNewDMs = taupfTausDiscriminationByDecayModeFindingNewDMs->at(i);
+            tau.ByMVA6VLooseElectronRejection = tauByMVA6VLooseElectronRejection->at(i);
+            tau.ByMVA6LooseElectronRejection = tauByMVA6LooseElectronRejection->at(i);
+            tau.ByMVA6MediumElectronRejection = tauByMVA6MediumElectronRejection->at(i);
+            tau.ByMVA6TightElectronRejection = tauByMVA6TightElectronRejection->at(i);
+            tau.ByMVA6VTightElectronRejection = tauByMVA6VTightElectronRejection->at(i);
+            tau.ByLooseMuonRejection3 = tauByLooseMuonRejection3->at(i);
+            tau.ByTightMuonRejection3 = tauByTightMuonRejection3->at(i);
+            tau.DecayMode = tauDecayMode->at(i);
+            tau.Charge = tauCharge->at(i);
+            tau.Dxy = tauDxy->at(i);
+            tau.dz = taudz->at(i);
+            tau.dxy = taudxy->at(i);
+            taus.push_back(tau);
+        }
     }
-  }
 
-  // sort by pT
-  std::sort(taus.begin(), taus.end(), [](Tau &p1, Tau &p2) -> bool { return p1.getPt() > p2.getPt(); });
-  nGoodTau = taus.size();
+    // sort by pT
+    std::sort(taus.begin(), taus.end(), [](Tau &p1, Tau &p2) -> bool { return p1.getPt() > p2.getPt(); });
+    nGoodTau = taus.size();
+}
+
+class Tau {
+    friend class Tau_Factory;
+
+   public:
+    Tau(Float_t _pt, Float_t _eta, Float_t _phi, Float_t _m) { this->p4.SetPtEtaPhiM(_pt, _eta, _phi, _m); }
+    Tau() { this->p4.SetPtEtaPhiE(0, 0, 0, 0); }
+
+    // getters
+    TLorentzVector getP4() { return p4; }
+    Float_t getPt() { return p4.Pt(); }
+    Float_t getEta() { return p4.Eta(); }
+    Float_t getPhi() { return p4.Phi(); }
+    Float_t getMass() { return p4.M(); }
+    Bool_t getIso(working_point);
+    Bool_t getDiscByDM(bool);
+    Bool_t getEleRejection(working_point);
+    Bool_t getMuRejection(working_point);
+    Bool_t getCombinedIsolationDeltaBetaCorr3Hits(working_point);
+    Int_t getDecayMode() { return DecayMode; }
+    Float_t getCharge() { return Charge; }
+    Float_t getDxy() { return Dxy; }
+    Float_t getDZ() { return dz; }
+    Float_t getDXY() { return dxy; }
+
+   private:
+    TLorentzVector p4;
+    Bool_t pass_vloose_iso, pass_loose_iso, pass_medium_iso, pass_tight_iso, pass_vtight_iso;
+
+    Bool_t pfTausDiscriminationByDecayModeFinding, pfTausDiscriminationByDecayModeFindingNewDMs, ByMVA6VLooseElectronRejection,
+        ByMVA6LooseElectronRejection, ByMVA6MediumElectronRejection, ByMVA6TightElectronRejection, ByMVA6VTightElectronRejection,
+        ByLooseMuonRejection3, ByTightMuonRejection3;
+
+    Int_t DecayMode;
+
+    Float_t Charge, Dxy, Mass, dz, dxy;
+};
+
+Bool_t Tau::getIso(working_point wp) {
+    if (wp == vloose) {
+        return pass_vloose_iso;
+    } else if (wp == loose) {
+        return pass_loose_iso;
+    } else if (wp == medium) {
+        return pass_medium_iso;
+    } else if (wp == tight) {
+        return pass_tight_iso;
+    } else if (wp == vtight) {
+        return pass_vtight_iso;
+    }
+    throw std::invalid_argument("Must use one of the provided tau isolation WPs");
+}
+
+Bool_t Tau::getDiscByDM(bool newDM = false) {
+    if (newDM) {
+        return pfTausDiscriminationByDecayModeFindingNewDMs;
+    }
+    return pfTausDiscriminationByDecayModeFinding;
+}
+
+Bool_t Tau::getEleRejection(working_point wp) {
+    if (wp == vloose) {
+        return ByMVA6VLooseElectronRejection;
+    } else if (wp == loose) {
+        return ByMVA6LooseElectronRejection;
+    } else if (wp == medium) {
+        return ByMVA6MediumElectronRejection;
+    } else if (wp == tight) {
+        return ByMVA6TightElectronRejection;
+    } else if (wp == vtight) {
+        return ByMVA6VTightElectronRejection;
+    }
+    throw std::invalid_argument("Must use one of the provided electron rejection WPs");
+}
+
+Bool_t Tau::getMuRejection(working_point wp) {
+    if (wp == loose) {
+        return ByLooseMuonRejection3;
+    } else if (wp == tight) {
+        return ByTightMuonRejection3;
+    }
+    throw std::invalid_argument("Muon rejection working point was neither Tight nor Loose");
 }
 
 #endif  // INTERFACE_TAU_FACTORY_H_
